Lots of updates
## preprocess.py

Commented out an environmental variable that was used in a local development environment that allows assignment process to specific GPU available on the system

## Dockerfile

Updated base image used for image running Kubeflow operators. There were conflicts between CUDA versions between the container and the base OS running on the hosts.

## train.py

1. Refactored a lot of the base code.
2. Updated directory creation logic
3. Added functionality to save model to disk (different from model checkpointing while training)
4. Added support for generating .json file on script execution for use in Kubeflow environment to enable Tesorboard 

## pipeline.py

Added capability for Kubeflow pipeline to request GPU on the underlying Kuberenetes infrastructure.